requirements.txt
python-dotenv==0.19.0
python-telegram-bot==13.14
requests==2.22.0
validate_email==1.3




